backend/pyproject.toml
[project]
name = "backbone-fastapi"
version = "3.1.0"
description = "Backbone FastAPI — Class-based views, DRF-style permissions, auto-admin, background tasks, cron scheduler, and email for MongoDB-backed FastAPI applications."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.20",
    # MongoDB / ODM (Beanie 2 + Motor; startup patch in backbone.compat.beanie_motor_aggregate)
    "beanie>=2.0.1",
    "motor>=3.7.1",
    # Pydantic & Settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.12.0",
    # Security
    "argon2-cffi>=25.1.0",
    "passlib>=1.7.4",
    "pyjwt>=2.11.0",
    # Email
    "jinja2>=3.1.6",
    "email-validator>=2.3.0",
    # Redis (cache + task queue)
    "redis>=5.0.1",
    # Slug generation
    "python-slugify>=8.0.4",
    # Scheduler
    "apscheduler>=3.10.4",
    # CLI
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    # HTTP client (for testing helpers)
    "httpx>=0.28.1",
    # File storage
    "cloudinary>=1.44.1",
    # PDF generation
    "reportlab>=4.2.2",
    "mongomock-motor>=0.0.36",
    "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mongomock-motor>=0.0.21",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pydantic>=2.0.0",
    "types-PyYAML",
]

[project.scripts]
backbone = "backbone.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backbone"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# ── Dev server ──────────────────────────────────────────────────────────────────
# uv run uvicorn main:app --reload
